requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
decorators==2.0.7
Django==5.1.4
django-admin-decorators==0.1
django-admin-list-filter-dropdown==1.0.3
django-ckeditor==6.7.2
django-filter==24.3
django-js-asset==3.0.1
django-render-partial==0.4
djangorestframework==3.16.0
fonttools==4.56.0
humanize==4.11.0
idna==3.10
joblib==1.4.2
kavenegar==1.1.2
kiwisolver==1.4.8
matplotlib==3.10.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.3.0
